# go_bootcamp/H8-Assign/docs/swagger.yaml
definitions:
  models.Car:
    properties:
      harga:
        type: integer
      merk:
        type: string
      pemilik:
        type: string
      typecars:
        type: string
    type: object
info:
  contact: {}
paths:
  /cars:
    post:
      consumes:
      - application/json
      description: Post details of car corresponding to the input id
      parameters:
      - description: create car
        in: body
        name: models.Car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Post details for a given
      tags:
      - cars
  /cars/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the order corresponding to the input id
      parameters:
      - description: ID of the car to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete car identified by the given id
      tags:
      - cars
    get:
      consumes:
      - application/json
      description: Get details of car corresponding to the input id
      parameters:
      - description: ID of the car
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Get details fro a given id
      tags:
      - cars
    patch:
      consumes:
      - application/json
      description: Update the car corresponding to the input Id
      parameters:
      - description: ID of the car to be updated
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Update car indetified by the given id
      tags:
      - cars
  /cars/getCar:
    get:
      consumes:
      - application/json
      description: Get details of all car
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Get details
      tags:
      - cars
swagger: "2.0"